backend sale update

+++ hello-shoes-backend/src/main/java/lk/ijse/gdse66/helloshoes/backend/dto/CustomerDTO.java
@@ -29,7 +29,7 @@ public class CustomerDTO implements Serializable {
     private String joinDate;
     private Address address;
     private Level level;
-    private Integer totalPoints;
+    private Double totalPoints;
     private Date dateOfBirth;
     private Timestamp recentPurchase;
